Move finders to /finders/dbfinder, /finders/rtfinder, /finders/gbfsfinder
