+++ finders/dbfinder/census_select.go
@@ -1,4 +1,4 @@
-package find
+package dbfinder
 
 import (
 	"strings"
