+++ finders/dbfinder/find_test.go
@@ -1,4 +1,4 @@
-package find
+package dbfinder
 
 import (
 	"testing"
