+++ finders/dbfinder/finder.go
@@ -1,4 +1,4 @@
-package find
+package dbfinder
 
 import (
 	"context"
