+++ finders/dbfinder/license_filter.go
@@ -1,4 +1,4 @@
-package find
+package dbfinder
 
 import (
 	"fmt"
