+++ finders/dbfinder/stop_time_select.go
@@ -1,4 +1,4 @@
-package find
+package dbfinder
 
 import (
 	"time"
