+++ finders/gbfsfinder/finder.go
