+++ finders/rtfinder/local.go
