+++ finders/rtfinder/lookup.go
