+++ finders/rtfinder/lookup_test.go
