+++ finders/rtfinder/redis_test.go
